fix compiler warning
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@3583 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ analyzer/pamfind.c
@@ -154,7 +154,7 @@ printHeader(FILE *       const ofP,
 
     fprintf(ofP, ")");
 
-    fprintf(ofP, "/%u", inpamP->maxval);
+    fprintf(ofP, "/%u", (unsigned)inpamP->maxval);
 
     fprintf(ofP, ":\n");
 }
